ch09_Project/smartfarm/requirements.txt
# Smart Farm Project Requirements
# python -m venv venv --system-site-packages

# Required Additional Packages
Flask==2.3.3
bluedot==2.0.0
Pillow==10.0.0
numpy==1.24.3
tflite-runtime==2.13.0
pygame==2.5.0
picamera2==0.3.12
adafruit-circuitpython-dht==4.1.0
gpiozero==1.6.2
RPLCD==1.3.0